pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terra4mice"
version = "0.1.0"
description = "State-Driven Development Framework - Terraform for your codebase"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Ultravioleta DAO", email = "ultravioletadao@gmail.com"}
]
keywords = [
    "development",
    "state-management",
    "terraform",
    "vivecoding",
    "specification",
    "tracking"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "mypy",
]

[project.scripts]
terra4mice = "terra4mice.cli:main"

[project.urls]
Homepage = "https://github.com/ultravioletadao/terra4mice"
Documentation = "https://github.com/ultravioletadao/terra4mice#readme"
Repository = "https://github.com/ultravioletadao/terra4mice"
Issues = "https://github.com/ultravioletadao/terra4mice/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
